moved code to registration_utils.py and wrote exception for alignment to track failures
